Create RequestSplitter. Extract methods from RequestHandler
